Step Five: Adding the logged-in user to a query
Adding to a query's where clause is a nice, scalable way to filter
results down to only those owned by the user.

Note that @PostAuthorize is still needed, if you want to throw an 403
instead of a 404, since once you add the user to the query, it might
return no results at all.